Add initial workflow for creating load balancer
This adds the action for creating a new load balancer. This only
includes the first panel of the workflow which allows creating the
load balancer object itself. Follow on patches will add support for
the listener, pool, monitor, and members.

This also adds the missing .eslintrc file and fixes code that was
failing the eslint rules.

There is a slight problem with the styling of the action button that
is probably an issue in the horizon action-list widget. This can be
treated as a bug and fixed later.

Depends-On: I230995c069d94ca41d70582b9c665289f0acb86e
Partially-Implements: blueprint horizon-lbaas-v2-ui
Change-Id: Ie0bfe6bd41a60259bc8642d4e046b420a50df55f