Do not allow deleting a listener that has a default pool
This change adds checks to the listener delete service to not allow deleting
a listener that currently has a default pool associated with it.  In the case
of the Delete batch action, an error message is displayed, while the delete
row action is simply just not displayed/available if a pool exists.

Partially-Implements: blueprint horizon-lbaas-v2-ui
Change-Id: Iddee0a5cee5ebbead9438a2add8ebcf57f69d348