fix unit tests
in log:
"xStatic libraries not found, please set up venv"
Added the post npm install step to setup the venv which
includes the static libraries and horizon.
And include that path in the base search paths for karma
and allow development to run tests as well.
Pull in test-shim.js from horizon master.

Closes-Bug: #1538264
Change-Id: Ia83a45f9e18335fa7802267d428d62b99a731dfe