Translation preparation for neutron-lbaas-dashboard
LBaaS v2 is now the only LBaaS implementation.
It is time to allow translator to translate LBaaS v2 dashboard.

This commit sets up related files according to the infra guide [1].
Also removes Babel related entries in setup.cfg. They are not
sufficient as Django requires two PO files (django and djangojs).

After this change is merged, I will propose the infra change.

This change can be part of newton release. If it is part of Newton
release of neutron-lbaas-dashboard, translators can translate it
even after the initial Newton release and we can import them
in the stable update release.

[1] http://docs.openstack.org/infra/manual/creators.html#enabling-translation-infrastructure

Change-Id: If3cf48dfe9c61499b2a4221dea956aa0bccf5f85