Add Constraints support
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove neutron-lbaas-dashboard.

Change-Id: Id0a31c878f39355bcb8b2e4b67b0c17b161c8603