Show the member status properties
This change adds the 'Operating Status' and 'Provisioning Status'
properties to the member attributes that are visible in the
members table and detail pages.  The member statuses are taken
from the Load Balancer status tree API, which returns all child
resources with their corresponding statuses for a particular Load
Balancer.

Change-Id: I2ff81940a62786b40efbfa9deed3e0dc972fb404
Closes-Bug: #1563444