Fix auth url for Barbican client
This patch fixes up the auth url if an invalid or no version prefix is
set in OPENSTACK_KEYSTONE_URL. An invalid prefix could be /v2.0 endpoint
when running Keystone v3.

keystoneclient.auth.token_endpoint is deprecated as of the 2.1.0 release
in favor of keystoneauth1.token_endpoint.Token. Reason why it is being
used here is because keystoneauth1 is not in our requirements.txt
(although it is a dependency pulled by required libraries) making this
patch not backportable if we were to add it now. A follow up patch
should handle this deprecation.

Story: 2005101
Task: 29723

Change-Id: Ib09c954180d4b545d05344dc2c318a48c9b8d4e4