Enable fail-fast on the gate queue
Zuul added fail-fast as a new option for project pipelines.
With this set, if any one of the voting jobs fails, zuul
will report a build failure immediately and abort the remaining
jobs.

This patch enables fail-fast on the gate pipeline as any one
failing job means the patch will not merge.

Change-Id: Ibcedfc15120a213f070e56ce514413fea911ac4d